refactor: create @pnpm/store-connection-manager
ref #2154
