+++ packages/pnpm/package.json
@@ -26,9 +26,7 @@
     "@pnpm/config": "workspace:5.0.2",
     "@pnpm/constants": "workspace:3.0.0",
     "@pnpm/core-loggers": "workspace:4.0.0",
-    "@pnpm/default-fetcher": "workspace:5.1.1",
     "@pnpm/default-reporter": "workspace:5.0.3",
-    "@pnpm/default-resolver": "workspace:5.0.2",
     "@pnpm/error": "workspace:1.0.0",
     "@pnpm/file-reporter": "0.1.0",
     "@pnpm/lifecycle": "workspace:8.0.0",
@@ -46,12 +44,12 @@
     "@pnpm/run-npm": "workspace:0.0.0",
     "@pnpm/semver-diff": "1.0.2",
     "@pnpm/server": "workspace:5.0.0",
+    "@pnpm/store-connection-manager": "workspace:0.0.0",
     "@pnpm/store-controller-types": "workspace:5.0.0",
     "@pnpm/store-path": "2.1.1",
     "@pnpm/types": "workspace:4.0.0",
     "@pnpm/utils": "workspace:0.12.0",
     "@types/archy": "0.0.31",
-    "@types/lru-cache": "^5.1.0",
     "@types/mz": "^2.7.0",
     "@types/wrap-ansi": "3.0.0",
     "@zkochan/libnpx": "11.0.2",
@@ -64,7 +62,6 @@
     "cross-spawn": "7.0.1",
     "delay": "4.3.0",
     "diable": "4.0.2",
-    "dir-is-case-sensitive": "1.0.2",
     "execa": "3.3.0",
     "find-packages": "workspace:7.0.0",
     "get-port": "5.0.0",
@@ -75,7 +72,6 @@
     "is-windows": "1.0.2",
     "load-json-file": "6.2.0",
     "loud-rejection": "2.2.0",
-    "lru-cache": "5.1.1",
     "make-dir": "3.0.0",
     "mem": "6.0.0",
     "mz": "2.7.0",
@@ -124,6 +120,7 @@
     "anonymous-npm-registry-client": "0.1.2",
     "byline": "5.0.0",
     "deep-require-cwd": "1.0.0",
+    "dir-is-case-sensitive": "1.0.2",
     "exists-link": "2.0.0",
     "ncp": "2.0.0",
     "normalize-newline": "3.0.0",
