+++ packages/pnpm/src/cmd/import.ts
@@ -2,13 +2,13 @@ import { docsUrl } from '@pnpm/cli-utils'
 import { WANTED_LOCKFILE } from '@pnpm/constants'
 import PnpmError from '@pnpm/error'
 import { readImporterManifestOnly } from '@pnpm/read-importer-manifest'
+import { createOrConnectStoreController } from '@pnpm/store-connection-manager'
 import rimraf = require('@zkochan/rimraf')
 import loadJsonFile = require('load-json-file')
 import path = require('path')
 import R = require('ramda')
 import renderHelp = require('render-help')
 import { install } from 'supi'
-import createStoreController from '../createStoreController'
 import { PnpmOptions } from '../types'
 
 export function types () {
@@ -36,7 +36,7 @@ export async function handler (
   const versionsByPackageNames = {}
   getAllVersionsByPackageNames(npmPackageLock, versionsByPackageNames)
   const preferredVersions = getPreferredVersions(versionsByPackageNames)
-  const store = await createStoreController(opts)
+  const store = await createOrConnectStoreController(opts)
   const installOpts = {
     ...opts,
     lockfileOnly: true,
