+++ packages/pnpm/src/cmd/server/stop.ts
@@ -1,13 +1,12 @@
 import { globalInfo, globalWarn } from '@pnpm/logger'
 import { connectStoreController } from '@pnpm/server'
+import { serverConnectionInfoDir, tryLoadServerJson } from '@pnpm/store-connection-manager'
 import storePath from '@pnpm/store-path'
 import delay from 'delay'
 import path = require('path')
 import processExists = require('process-exists')
 import killcb = require('tree-kill')
 import { promisify } from 'util'
-import { tryLoadServerJson } from '../../createStoreController'
-import serverConnectionInfoDir from '../../serverConnectionInfoDir'
 
 const kill = promisify(killcb) as (pid: number, signal: string) => Promise<void>
