+++ packages/store-connection-manager/README.md
@@ -1,0 +1,15 @@
+# @pnpm/store-connection-manager
+
+> Create a direct pnpm store controller or connect to a running store server
+
+[![npm version](https://img.shields.io/npm/v/@pnpm/store-connection-manager.svg)](https://www.npmjs.com/package/@pnpm/store-connection-manager)
+
+## Installation
+
+```sh
+<pnpm|npm|yarn> add @pnpm/store-connection-manager
+```
+
+## License
+
+MIT © [Zoltan Kochan](https://www.kochan.io/)
