+++ packages/store-connection-manager/package.json
@@ -1,0 +1,57 @@
+{
+  "name": "@pnpm/store-connection-manager",
+  "version": "0.0.0",
+  "description": "Create a direct pnpm store controller or connect to a running store server",
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "files": [
+    "lib"
+  ],
+  "engines": {
+    "node": ">=10"
+  },
+  "scripts": {
+    "lint": "tslint -c tslint.json src/**/*.ts test/**/*.ts",
+    "tsc": "rimraf lib && tsc",
+    "test": "pnpm run tsc && pnpm run lint",
+    "prepublishOnly": "pnpm run tsc"
+  },
+  "repository": "https://github.com/pnpm/pnpm/blob/master/packages/store-connection-manager",
+  "keywords": [
+    "pnpm"
+  ],
+  "author": "Zoltan Kochan <z@kochan.io> (https://www.kochan.io/)",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/pnpm/pnpm/issues"
+  },
+  "homepage": "https://pnpm.js.org",
+  "devDependencies": {
+    "@pnpm/logger": "3.1.0",
+    "@pnpm/store-connection-manager": "link:",
+    "@types/lru-cache": "5.1.0",
+    "@types/mz": "^2.7.0",
+    "pnpm": "link:../pnpm",
+    "rimraf": "3.0.0"
+  },
+  "dependencies": {
+    "@pnpm/cli-utils": "workspace:0.0.0",
+    "@pnpm/config": "workspace:5.0.2",
+    "@pnpm/default-fetcher": "workspace:5.1.1",
+    "@pnpm/default-resolver": "workspace:5.0.2",
+    "@pnpm/error": "workspace:1.0.0",
+    "@pnpm/package-store": "workspace:6.0.1",
+    "@pnpm/server": "workspace:5.0.0",
+    "@pnpm/store-path": "2.1.1",
+    "delay": "4.3.0",
+    "diable": "4.0.2",
+    "dir-is-case-sensitive": "1.0.2",
+    "lru-cache": "5.1.1",
+    "make-dir": "3.0.0",
+    "mz": "2.7.0"
+  },
+  "peerDependencies": {
+    "@pnpm/logger": "^3.1.0",
+    "pnpm": "*"
+  }
+}
