+++ packages/store-connection-manager/src/createNewStoreController.ts
