+++ packages/store-connection-manager/src/serverConnectionInfoDir.ts
